fix dtype again (?)

+++ ultralytics/nn/modules/block.py
@@ -2139,10 +2139,14 @@ class FeatureSelectionModule(nn.Module):
         device = raw_preds.device
 
         # Initialize tensors (same as threshold method)
-        selected_cls_features = torch.zeros(batch_size, self.topk_post, vid_features.shape[-1], device=device)
-        selected_reg_features = torch.zeros(batch_size, self.topk_post, reg_features.shape[-1], device=device)
-        selected_boxes = torch.zeros(batch_size, self.topk_post, 4, device=device)
-        selected_scores = torch.zeros(batch_size, self.topk_post, device=device)
+        selected_cls_features = torch.zeros(
+            batch_size, self.topk_post, vid_features.shape[-1], device=device, dtype=raw_preds.dtype
+        )
+        selected_reg_features = torch.zeros(
+            batch_size, self.topk_post, reg_features.shape[-1], device=device, dtype=raw_preds.dtype
+        )
+        selected_boxes = torch.zeros(batch_size, self.topk_post, 4, device=device, dtype=raw_preds.dtype)
+        selected_scores = torch.zeros(batch_size, self.topk_post, device=device, dtype=raw_preds.dtype)
         selected_indices = torch.zeros(batch_size, self.topk_post, dtype=torch.long, device=device)
         if not self.training and batch_size > 1:
             breakpoint()
